Added Python Cheet Sheet
This sheet is everything you need to begin coding independently. Use it as a resource when you forget a particular method or syntax. From https://ehmatthes.github.io/pcc/cheatsheets/README.html

Be sure to check out the book it is based on https://www.nostarch.com/pythoncrashcourse
I am not affiliated with them but it is an excellent book that I highly recommend.